+++ modules/sugar/src/main/ts/ephox/sugar/api/view/Width.ts
@@ -1,4 +1,5 @@
 import { Dimension } from '../../impl/Dimension';
+import * as RuntimeSize from '../../impl/RuntimeSize';
 import { SugarElement } from '../node/SugarElement';
 import * as Css from '../properties/Css';
 
@@ -13,6 +14,10 @@ const get = (element: SugarElement<HTMLElement>): number => api.get(element);
 
 const getOuter = (element: SugarElement<HTMLElement>): number => api.getOuter(element);
 
+const getInner = RuntimeSize.getInnerWidth;
+
+const getRuntime = RuntimeSize.getWidth;
+
 const setMax = (element: SugarElement<HTMLElement>, value: number): void => {
   // These properties affect the absolute max-height, they are not counted natively, we want to include these properties.
   const inclusions = [ 'margin-left', 'border-left-width', 'padding-left', 'padding-right', 'border-right-width', 'margin-right' ];
@@ -23,6 +28,8 @@ const setMax = (element: SugarElement<HTMLElement>, value: number): void => {
 export {
   set,
   get,
+  getInner,
   getOuter,
+  getRuntime,
   setMax
 };