+++ modules/tinymce/src/plugins/image/test/ts/browser/ImagePluginTest.ts
@@ -11,9 +11,16 @@ import { advancedTabSelectors, assertInputValue, fillActiveDialog, ImageDialogDa
 describe('browser.tinymce.plugins.image.ImagePluginTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     plugins: 'image',
+    toolbar: 'image',
     indent: false,
-    base_url: '/project/tinymce/js/tinymce'
-  }, [ Plugin, Theme ]);
+    base_url: '/project/tinymce/js/tinymce',
+    setup: (editor) => {
+      editor.ui.registry.addContextToolbar('test-image', {
+        predicate: (node) => node.nodeName.toLowerCase() === 'img',
+        items: 'image'
+      });
+    }
+  }, [ Plugin, Theme ], true);
 
   const pInitAndOpenDialog = async (editor: Editor, content: string, cursorPos: Cursors.CursorSpec | Cursors.RangeSpec) => {
     editor.settings.image_advtab = true;
@@ -218,4 +225,16 @@ describe('browser.tinymce.plugins.image.ImagePluginTest', () => {
       }
     )
   );
+
+  it('TINY-3463: Ensure initial toolbar button state shows correctly', async () => {
+    const editor = hook.editor();
+    editor.setContent('<p>Content</p><p><img src="image.png"></p>');
+
+    TinySelections.setCursor(editor, [ 0, 0 ], 4);
+    await TinyUiActions.pWaitForUi(editor, '.tox-tbtn[title="Insert/edit image"]:not(.tox-tbtn--enabled)');
+
+    TinySelections.setSelection(editor, [ 1 ], 0, [ 1 ], 1);
+    await TinyUiActions.pWaitForUi(editor, '.tox-tbtn.tox-tbtn--enabled[title="Insert/edit image"]');
+    await TinyUiActions.pWaitForUi(editor, '.tox-pop .tox-tbtn.tox-tbtn--enabled[title="Insert/edit image"]');
+  });
 });