+++ modules/tinymce/src/plugins/media/main/ts/core/Selection.ts
@@ -12,6 +12,9 @@ import * as UpdateHtml from './UpdateHtml';
 declare let escape: any;
 declare let unescape: any;
 
+const isMediaElement = (element: Element): boolean =>
+  element.hasAttribute('data-mce-object') || element.hasAttribute('data-ephox-embed-iri');
+
 const setup = (editor: Editor): void => {
   editor.on('click keyup touchend', () => {
     const selectedNode = editor.selection.getNode();
@@ -50,5 +53,6 @@ const setup = (editor: Editor): void => {
 };
 
 export {
-  setup
+  setup,
+  isMediaElement
 };