+++ modules/darwin/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Fixed
+- Some of the selection annotation attributes were not cleared if the `Ephemera.selectedSelector` attribute was not present.
+
 ## 7.0.0 - 2021-08-26
 
 ## Added