+++ modules/sugar/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added new `parentElement` function to the `Traverse` API.
 - Added new `setOptions` function to the `Attribute` API.
+- Added new `getInner` and `getRuntime` functions to the `Width` and `Height` APIs.
 
 ### Fixed
 - Disabled `window.visualViewport` in Mozilla Firefox as it was returning an incorrect value for `pageTop` when using `position: 'fixed'`.